fix(development-harness): scope startup-sync gate fixture to whole plugin subtree
_disable_startup_sync only lived in tests/conftest.py, so it never applied to
sibling directories like tests_backlog/. Tests there entered the real
_backlog_lifespan and paid a ~90-109s teardown cost per test waiting on the
live GitHub sync task (tests_backlog/test_artifact_provider.py:
347s -> 13s for the full file).

Move the fixture to the plugin-root conftest.py so it covers every test
directory in the subtree, with a new allow_startup_sync marker opt-out for
the four backlog_core/tests/test_startup_sync.py tests that assert on the
real gate. Keep the backlog_core.server import lazy inside the fixture
(matching the original design) rather than top-level, since a module-level
import would make the FastMCP tool-surface import a mandatory cost for
every nested pytest subprocess that collects this rootdir conftest,
including test_network_guard.py's tightly-timed guard probes.

Follow-up (deliberately out of scope): backlog item #2930 will add
tests_backlog/ to testpaths now that the teardown cost is fixed.
